+++ src/app/chatbot-design-studio/utils.ts
@@ -201,7 +201,7 @@ export const ELEMENTS_LIST = [
     type: TYPE_ACTION.EMAIL,
     src: "assets/cds/images/actions/send_email.svg",
     description:
-      'This action send an email to the specified users group or email addresses.<br>You can use a comma sepatated addresses list.<br>i.e. “andrea@tiledesk.com, gab@tiledesk.com"<br>You can use the special tag “@everyone” to send an email to each of the GPT project teamates.<br><br>You can also use the name of a single user group using the group name. i.e. “sales”',
+      'This action send an email to the specified users group or email addresses.<br>You can use a comma sepatated addresses list.<br>i.e. “andrea@tiledesk.com, gab@tiledesk.com"<br>You can use the special tag “@everyone” to send an email to each of the ChatRep project teamates.<br><br>You can also use the name of a single user group using the group name. i.e. “sales”',
   },
   {
     name: "WhatsApp Static",
@@ -344,7 +344,7 @@ export const ACTIONS_LIST = {
     type: TYPE_ACTION.EMAIL,
     src: "assets/cds/images/actions/send_email.svg",
     description:
-      'This action send an email to the specified users group or email addresses.<br>You can use a comma sepatated addresses list.<br>i.e. “andrea@tiledesk.com, gab@tiledesk.com"<br>You can use the special tag “@everyone” to send an email to each of the GPT project teamates.<br><br>You can also use the name of a single user group using the group name. i.e. “sales”',
+      'This action send an email to the specified users group or email addresses.<br>You can use a comma sepatated addresses list.<br>i.e. “andrea@tiledesk.com, gab@tiledesk.com"<br>You can use the special tag “@everyone” to send an email to each of the ChatRep project teamates.<br><br>You can also use the name of a single user group using the group name. i.e. “sales”',
   },
   WHATSAPP_STATIC: {
     name: "WhatsApp Static",