+++ src/app/knowledge-bases/modals/modal-preview-knowledge-base/modal-preview-knowledge-base.component.ts
@@ -26,10 +26,10 @@ export class ModalPreviewKnowledgeBaseComponent implements OnInit {
   context: string;
 
   // models_list = [
-  //   { name: "GPT-3.5 Turbo (ChatGPT)", value: "gpt-3.5-turbo" }, 
-  //   { name: "GPT-4 (ChatGPT)", value: "gpt-4" },
-  //   { name: "GPT-4 Turbo Preview (ChatGPT)", value: "gpt-4-turbo-preview" }, 
-  //   { name: "GPT-4o (ChatGPT)", value: "gpt-4o" }
+  //   { name: "ChatRep-3.5 Turbo (ChatGPT)", value: "gpt-3.5-turbo" }, 
+  //   { name: "ChatRep-4 (ChatGPT)", value: "gpt-4" },
+  //   { name: "ChatRep-4 Turbo Preview (ChatGPT)", value: "gpt-4-turbo-preview" }, 
+  //   { name: "ChatRep-4o (ChatGPT)", value: "gpt-4o" }
   // ];
   // selectedModel: any = this.models_list[1].value;