+++ src/app/utils/util.ts
@@ -405,7 +405,7 @@ export const featuresPlanC = [
   "Dialogflow connector",
   "Rasa connector",
   "SMTP Settings",
-  "Support to host GPT on your Infrastructure",
+  "Support to host ChatRep on your Infrastructure",
   "Premium Customer Support",
 ];
 
@@ -594,7 +594,7 @@ export const featuresPlanF = [
   "Connector with 3rd party AI",
   "Automations Log",
   "SMTP Settings",
-  "Support to host GPT on your Infrastructure",
+  "Support to host ChatRep on your Infrastructure",
   "Premium Customer Support",
 ];
 
@@ -748,31 +748,31 @@ export function loadTokenMultiplier(ai_models) {
 }
 
 // export const TYPE_GPT_MODEL = {
-//     'GPT-3': { name: "GPT-3 (DaVinci)", value: "text-davinci-003", status: "inactive"},
-//     'GPT-3.5' : { name: "GPT-3.5 Turbo (ChatGPT)", value: "gpt-3.5-turbo", status: "active"},
-//     'GPT-4' : { name: "GPT-4 (ChatGPT)", value: "gpt-4", status: "active"},
-//     'GPT-4-turbo-preview': { name: "GPT-4 Turbo Preview (ChatGPT)", value: "gpt-4-turbo-preview", status: "active"},
-//     'GPT-4o': { name: "GPT-4o (ChatGPT)", value: "gpt-4o", status: "active"}
+//     'ChatRep-3': { name: "ChatRep-3 (DaVinci)", value: "text-davinci-003", status: "inactive"},
+//     'ChatRep-3.5' : { name: "ChatRep-3.5 Turbo (ChatGPT)", value: "gpt-3.5-turbo", status: "active"},
+//     'ChatRep-4' : { name: "ChatRep-4 (ChatGPT)", value: "gpt-4", status: "active"},
+//     'ChatRep-4-turbo-preview': { name: "ChatRep-4 Turbo Preview (ChatGPT)", value: "gpt-4-turbo-preview", status: "active"},
+//     'ChatRep-4o': { name: "ChatRep-4o (ChatGPT)", value: "gpt-4o", status: "active"}
 // }
 
 export const TYPE_GPT_MODEL = {
-  "GPT-3": {
-    name: "GPT-3 (DaVinci)",
+  "ChatRep-3": {
+    name: "ChatRep-3 (DaVinci)",
     value: "text-davinci-003",
     status: "inactive",
   },
-  "GPT-3.5": {
-    name: "GPT-3.5 Turbo",
+  "ChatRep-3.5": {
+    name: "ChatRep-3.5 Turbo",
     value: "gpt-3.5-turbo",
     status: "active",
   },
-  "GPT-4": { name: "GPT-4", value: "gpt-4", status: "active" },
-  "GPT-4-turbo-preview": {
-    name: "GPT-4 Turbo",
+  "ChatRep-4": { name: "ChatRep-4", value: "gpt-4", status: "active" },
+  "ChatRep-4-turbo-preview": {
+    name: "ChatRep-4 Turbo",
     value: "gpt-4-turbo-preview",
     status: "active",
   },
-  "GPT-4o": { name: "GPT-4o", value: "gpt-4o", status: "active" },
+  "ChatRep-4o": { name: "ChatRep-4o", value: "gpt-4o", status: "active" },
 };
 
 export const CHANNELS_NAME = {