+++ src/app/widget_components/widget-set-up/widget-set-up-base/widget-set-up-base.component.ts
@@ -61,7 +61,7 @@ export class WidgetSetUpBaseComponent
       },
     ],
     en: {
-      wellcomeTitle: "Hi, welcome to GPT 👋 ",
+      wellcomeTitle: "Hi, welcome to ChatRep 👋 ",
       wellcomeMsg: "How can we help you?",
       calloutTitle: "Need Help?",
       calloutMsg: "Click here and start chatting with us!",
@@ -72,7 +72,7 @@ export class WidgetSetUpBaseComponent
         "🤔 All operators are offline at the moment. You can anyway describe your problem. It will be assigned to the support team who will answer you as soon as possible.",
     },
     it: {
-      wellcomeTitle: "Ciao, benvenuto su GPT 👋",
+      wellcomeTitle: "Ciao, benvenuto su ChatRep 👋",
       wellcomeMsg: "Come possiamo aiutarti?",
       calloutTitle: "Bisogno di aiuto?",
       calloutMsg: "Clicca qui e inizia a chattare con noi!",