+++ src/app/ws_requests/ws-requests-list/ws-requests-list.component.ts
@@ -338,7 +338,7 @@ export class WsRequestsListComponent
     );
     let baseUrl = this.CHAT_BASE_URL + "#/conversation-detail/";
     let url = baseUrl;
-    const myWindow = window.open(url, "_self", "GPT - Open Source Live Chat");
+    const myWindow = window.open(url, "_self", "ChatRep - Open Source Live Chat");
     myWindow.focus();
   }